# Move away from using KeyValuePair for entry storage.